requirements.txt
# Django ve Core
Django==5.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0

# Channels ve WebSocket
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Redis ve Celery
redis==5.0.1
celery==5.3.6

# Database
psycopg2-binary==2.9.9

# AWS S3 (opsiyonel)
boto3==1.34.69

# Rate Limiting
django-ratelimit==4.1.0

# Environment Variables
python-dotenv==1.0.1

# Image Processing
Pillow==11.2.1

# Cryptography
cryptography==45.0.4

# Development Tools
django-debug-toolbar==4.3.0
coverage==7.4.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0 